// gen-ts/api.ts
export interface AuthResponse {
  /**
   *
   * @type {string}
   * @memberof AuthResponse
   */
  token?: string
  /**
   *
   * @type {string}
   * @memberof AuthResponse
   */
  type?: string
  /**
   *
   * @type {string}
   * @memberof AuthResponse
   */
  id?: string
  /**
   *
   * @type {string}
   * @memberof AuthResponse
   */
  email?: string
  /**
   *
   * @type {string}
   * @memberof AuthResponse
   */
  role?: string
}
// ...
export const Role = {
  Admin: 'ADMIN',
  WarehouseWorker: 'WAREHOUSE_WORKER',
  Employee: 'EMPLOYEE',
  Administration: 'ADMINISTRATION',
} as const

export type Role = (typeof Role)[keyof typeof Role]